Merge Garaj and TP

+++ include/LuaInterface/LuaInterface.h
@@ -67,7 +67,7 @@ public:
 	     * \param name  name of variable
 	     * \return int  value of given variable
 	     */
-	long int getInt( QString name );
+	int getInt( QString name );
 
 	/**
 	     * \fn getInt
@@ -77,7 +77,7 @@ public:
 	     * \param args      names of table and indexes to access nested variable
 	     * \return int  value of given variable
 	     */
-	long int getInt( unsigned int length, QString args[] );
+	int getInt( unsigned int length, QString args[] );
 
 	/**
 	     * \fn getDouble
