+++ src/OsgQtBrowser/QGraphicsViewAdapter.cpp
@@ -405,7 +405,7 @@ void QGraphicsViewAdapter::render()
 	// paint the image with the graphics view
 	QPainter painter( &image );
 	QRectF destinationRect( 0, 0, image.width(), image.height() );
-	QRect sourceRect( 0, 0, _graphicsScene->width(), _graphicsScene->height() );
+	QRect sourceRect( 0, 0, ( int )( _graphicsScene->width() ), ( int )( _graphicsScene->height() ) );
 	_graphicsView->render( &painter, destinationRect, sourceRect );
 	painter.end();
 #else
