+++ src/Util/ApplicationConfig.cpp
@@ -2,6 +2,7 @@
 
 #include "OpenThreads/Mutex"
 
+#include <QDebug>
 #include <QFile>
 #include <QDebug>
 
@@ -43,11 +44,11 @@ QString Util::ApplicationConfig::getValue( QString key )
 }
 
 
-long Util::ApplicationConfig::getNumericValue(
+int Util::ApplicationConfig::getNumericValue(
 	QString key,
-	std::auto_ptr<long> minValue,
-	std::auto_ptr<long> maxValue,
-	const long defaultValue
+	std::auto_ptr<int> minValue,
+	std::auto_ptr<int> maxValue,
+	const int defaultValue
 )
 {
 	if ( !_map.contains( key ) ) {
@@ -56,7 +57,7 @@ long Util::ApplicationConfig::getNumericValue(
 
 	QString value = getValue( key );
 	bool ok = true;
-	long result = value.toLong( &ok );
+	int result = value.toInt( &ok );
 
 	if ( !ok ) {
 		return defaultValue;
